Refactor docs about migrating to Patroni
This commit is an attempt of refactoring the docs about migrating
to Patroni.

The original docs mentioned the steps could only be performed in
a single node cluster, while in fact in can convert a multi-node
cluster to Patroni.

This commit also adds some information about how to deal with
existing slots, and enhances information about creation of users.

References: PAT-143.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>