Refactor Connection class
1. stop using the same cursor all the time, it creates problems when not
   carefully used from different threads.
2. introduce query() method in the Connection class and make it return a
   result set when it is possible.
3. refactor most of the code that is relying (directly or indirectly) on
   the Connection object to use the query() method as much as possible.

This refactoring helps with reducing code complexity and will help with
future introduction of a separate database connection for the REST API
thread. The last one will help to improve reliability when system is
under significant stress when simple monitoring queries are taking
seconds to execute and the REST API starts blocking the main thread.