Fix `IntValidator` regarding validation of value `0`
Previous to this commit `IntValidator` would always consider the
value `0` invalid, even if in the allowed range.

The problem was that `parse_int` was returning `0` in the following
line:

```python
value = parse_int(value, self.base_unit) or ""
```

However the `or ""` was evaluating to an empty string.

As `parse_int` returns either an `int` if able to parse, or `None`
otherwise, the `isinstance(value, int)` is enough to error out
when not a valid `int`.

Closes #2817

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>