Fix bug with kubernetes.standby_leader_label_value
When running with the leader lock Patroni was just setting the `role`
label to `master` and effectively kubernetes.standby_leader_label_value
feature never worked.

Now it is fixed, but in order to not introduce breaking changes we just
update default value of the `standby_leader_label_value` to the
`master`.