Document how to change Postgres settings that touch shared memory
Some special handling is required when changing either of these
settings in a Postgres cluster that has standby nodes:

* `max_connections`
* `max_prepared_transactions`
* `max_locks_per_transaction`
* `max_wal_senders`
* `max_worker_processes`

If one attempts to decrease `max_connections` dynamic setting and
restart all nodes at the same time (primary and standbys), Patroni
will refuse to apply the new value on the standbys and require the
user to restart it again later, once replication catches up.

That behavior is correct, but it is not documented.

This commit adds information to documentation about that behavior
and why it's required.

References: PAT-166.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>