`patronictl --help` was showing `ctl` function's docstring
`patronictl` is implemented using `click` module, and that module
uses the functions' docstrings for creating a helper text.

As a consequence the docstring fo `ctl` function was being shown
to the user, which doesn't make sense.

this commit fixes that issue by Adding a user-friendly description
to be shown on `patronictl --help`. We use a `\f` to tell `click`
when to stop capturing text to show in the helper.

Note that `patronictl` commands are implemented using `@ctl.command`
decorator, and we always provide them with `help` argument. That said,
none of the subcommands are affected by the aforementioned issue,
only the entry point of the CLI.

References: PAT-201.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>