Initial work on documentation of `patronictl`
This commit is an initial work on the documentation of ``patronictl``.

So far we basically added a section called ``patronictl`` to the
left menu.

Inside that section we created a page with this structure:

- `patronictl`: describes what it is
    - `Configuraiton`: how to configure `patronictl`
    - `Usage`: how to use the CLI
        - `patronictl dsn`: describe how to use `patronictl dsn` command.

We adopted a top-down approach when writing this document. We start
by describing the outer most parts, and then keep writing new sections
that specialize the knowledge.

References: PAT-200.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>