Permanent physical slots on standby nodes
Create permanent physical replication slots on standby nodes and use
`pg_replication_slot_advance()` function to move them forward.

The `restart_lsn` is advanced based on values stored in the `/status`
key by the primary node.

When slot is created on a replica it could be ahead the same slot on the
primary and therefore there is some period of time when it doesn't
protect WAL files from being recycled.