Fix bug in `patronictl query` command
Previous to this commit `patronictl query` was working only if `-r`
argument was provided to the command. Otherwise it would face issues:

* If neither `-r` nor `-m` were provided:

```
 PGPASSWORD=zalando patronictl -c postgres0.yml query -U postgres -c "SHOW PORT"
2023-09-12 17:45:38	No connection to role=None is available
```

* If only `-m` was provided:

```
$ PGPASSWORD=zalando patronictl -c postgres0.yml query -U postgres -c "SHOW PORT" -m postgresql0
2023-09-12 17:46:15	No connection to member postgresql0 is available
```

This issue was a regression introduced by `4c3e0b9382820524239d2aa4d6b95379ef1291db`
through PR #2687.

Through that PR we decided to move the common logic used to check
mutually exclusiveness of `--role` and `--member` arguments to `get_any_member`
function.

However, previous to that change `role` variable would assume the default value
of `any` in `query` method, before `get_any_member` was called, which was
not the case after the change.

This commit fixes that issue by adding a handler in `get_cursor` function
to `role=None`. As `role` defaulting to `any` is handled in a sub-call to
`get_any_member`, we are apparently safe in `get_cursor` to return the
cursor if `role=None`.

Unit tests were updated accordingly.

References: PAT-204.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>