Read GUC's values when joining running Postgres
If restarted in pause Patroni was discarding `synchronous_standby_names`
from `postgresql.conf` because in the internal cache this values was set
to `None`. As a result synchronous replication transitioned to a broken
state, with no synchronous replicas according to the
`synchronous_standby_names` and Patroni not selecting/setting the new
synchronous replicas (another bug).

To solve the problem of broken initial state and to avoid similar issues
with other GUC's we will read GUC's value if Patroni is joining running
Postgres.