Detect and solve inconsistency between /sync and actual sync nodes
Patroni is changing `synchronous_standby_names` and the `/sync` key in a
very specific order, first we add nodes to `synchronous_standby_names`
and only after, when they are recognized as synchronous they are added
to the `/sync` key. When removing nodes the order is different: they are
first removed from the `/sync` key and only after that from the
`synchronous_standby_names`.

As a result Patroni expects that either actual synchronous nodes will
match with the nodes listed in the `/sync` key or that new candidates to
synchronous nodes will not match with nodes listed in the `/sync` key.
In case if `synchronous_standby_names` was removed from the
`postgresql.conf`, manually, or due the the bug (#2876), the state
becomes inconsistent because of the wrong order of updates.

To solve inconsistent state we introduce additional checks and will
update the `/sync` key with actual names of synchronous nodes (usually
empty set).