Take into account current role when deciding on removal of member ZNode
Patroni doesn't watch on all changes of member keys in order to not
create too much load on ZooKeeper, but only subscribes to changes (ZNodes
added or deleted) in the `/member` directory. Therefore when some
important fields in the value are updated we remove and recreate ZNode in
order to notify the leader or other members.

The leader should remove the member key only when the
`checkpoint_after_promote` value is changed and replicas when the `state` is
changed to/from `running`.

We don't care about the `version` field, because Patroni version can't
be changed without restart, what will case ZooKeeper `session_id` to
change it anyway.

This fix hopefully will reduce failures of behave tests on GH Actions.