Set citus.local_hostname
There are cases when Citus wants to have a connection to the local
postgres. By default it uses `localhost` for that, which is not alwasy
available. To solve it we will set `citus.local_hostname` GUC to custom
value, which is the same as Patroni uses to connect to Postgres.