Handle AuthOldRevision error
If retries are requested - try to get a new new token and repeat the
request. Repeat it in a loop until request is successfully executed or
until `retry_timeout` is exhausted. This is the only way of solving a
race condition, because between authentication and executing the request
yet another modification of the user database in Etcd might happen.

In case if the request doesn't have to be immediately retried - set a
flag that the next API request should perform the authentication first
and let Patroni to naturally repeat the request on the next heartbeat
loop.

Co-authored-by: Kenny Do <kedo@render.com>