Add support for additional parameters on custom bootstrap
Previous to this commit, if a user would ever like to add parameters
to the custom bootstrap script call, they would need to configure
Patroni like this:

```
bootstrap:
  method: custom_method_name
  custom_method_name:
    command: /path/to/my/custom_script --arg1=value1 --arg2=value2 ...
```

This commit changes that behavior so we achieve a similar behavior
that is seen when using `create_replica_methods`, i.e., we allow
the following syntax:

```
bootstrap:
  method: custom_method_name
  custom_method_name:
    command: /path/to/my/custom_script
    arg1: value1
    arg2: value2
```

All keys in the mapping which are not recognized by Patroni, will
be dealt with as if they were additional named arguments to be
passed down to the `command` call.

References: PAT-218.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>