Modify behavior tests to test the just introduced feature
Through the previous commit we added the capability of specifying
arguments for the `command` of a custom bootstrap as a mapping in
the YAML file.

This commit adds behave testing for that capability. We introduce
a new fake argument in the `backup_restore.py` helper script just
to check that the custom keys in the YAML mapping are passed to
the called script.

References: PAT-218.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>