Add a contrib script for remote Barman recovery
This commit adds a contrib script, which can be used as a custom
bootstrap method, or as a custom create replica method.

The script communicates with the `pg-backup-api` on the Barman node
so Patroni is able to restore a Barman backup remotely.

The `--help` option of the script, along with the script docstring,
should provide enough context on how to use fill its parameters.

References: PAT-216.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>