Apply suggestions from code review -- add retry mechanism
Previous to this commit the `barman_recover.py` script was quitting
on the very first malformed response received from the ``pg-backup-api``.

This commit implements a retry mechanism which can be configured by
the user in terms of:

* Wait between attempts
* Maximum number of attempts

This should make the scrip more resilient to temporary failures on
the internet that could create a malformed response body.

As a side note, we also modified the purpose of the SSL connections
from `CLIENT_AUTH` to `SERVER_AUTH` as we are opening a client
socket.

References: PAT-216.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>