Verify that replica nodes received checkpoint LSN on shutdown
In case if archiving is enabled the `latest_checkpoint_location()`
method returns LSN of the prev (SWITCH) record, which points to the
beginning of the WAL file. It is done in order to make it possible to
safely promote replica which recovers WAL files from the archive and
wasn't streaming when the primary was stopped.

But, in certain cases using the LSN pointing to SWITCH record was
causing unnecessary pg_rewind, if replica didn't managed to replay tha
shutdown checkpoint record before it was promoted.

In order to mitigate the problem we need to check that replica
received/replayed exactly the shutdown checkpoint LSN. But, at the same
time we will still write SWITCH record LSN to the `/status` key when
releasing the leader lock.