PoC: Abstract CitusHandler and decouple it from configuration
the main issue was that the configuration for Citus handler and for DCS
existed in two places, while ideally AbstractDCS should not know many
details about what kind of MPP is in use.

To solve the problem we first dynamically create ne object implementing
AbstractMPP interfaces, which is a configuration for DCS. Later this
object is used to instantiate the class implementing AbstractMPPHandler
interfaces.

This is just a starting point, which does some heavy lifting. As a next
steps the following should be done:
1. AbstractMPP and AbstractMPPHandler and get_mpp() should be moved to
   patroni/postgres/mpp/__init__.py
2. Null and NullHandler should be moved to patroni/postgres/mpp/null.py
3. patroni/postgres/citus.py renamed to patroni/postgres/mpp/citus.py
4. get_mpp() should be working similar to get_dcs()
5. All kind of variables named after Citus in files different from
   patroni/postgres/mpp/citus.py should be renamed.

In other words this commit takes over the most complex part of #2940,
which was never implemented.