Abstract CitusHandler and decouple it from configuration
The main issue was that the configuration for Citus handler and for DCS
existed in two places, while ideally AbstractDCS should not know many
details about what kind of MPP is in use.

To solve the problem we first dynamically create ne object implementing
AbstractMPP interfaces, which is a configuration for DCS. Later this
object is used to instantiate the class implementing AbstractMPPHandler
interfaces.

Co-authored-by: Junwang Zhao <zhjwpku@gmail.com>