Use consistent read when fetching just updated sync key
Consul doesn't provide any interface to immediately get `ModifyIndex`
for the key that we just updated, therefore we have to perform an
explicit read operation. By default stale reads are allowed and
sometimes we may read stale data. As a result write_sync_state() call
was considered as failed. To mitigate the problem we switch to
`consistent` reads when that executed after update of the `/sync` key.

Close #2972