Create citus database and extension idempotently
Consider a task: we want to create an extension _before_ citus in a
database. Currently `post_bootstrab` script is executed before
`CitusHandler.bootstrap()` method, which seems to allow doing that, but
in fact `CitusHandler.bootstrap()` will fail to create already existing
database and as a result the whole bootstrap will fail.

Changing the order of execution of `post_bootstrab` hook and
`CitusHandler.bootstrap()` seems to be useless, because it will not
allow creating another extension _before_ citus. Therefore the only way
of solving it is making CREATE DATABASE and CREATE EXTENSION idempotent.
It will allow to create citus database and all dependencies from the
`post_bootstrab` hook.