Close the doors for a possible future bug in the config generator
The `AbstractConfigGenerator._format_config` method was missing a
comma in the declaration of a tuple. As a consequence it was
concatenating the strings `ctl` and `citus` instead of creating two
separate items in the tuple.

There is currently no observed bug from that issue in the code
because the template configuration created by the method
`AbstractConfigGenerator.get_template_config` doesn't include
either of `ctl` or `citus` keys.

However, it is still important that we close the doors for possible
future bugs that would come up if we ever attempt to use either
of those keys in the template, for example.

References: PAT-231.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>