Add `config-switch` sub-command to `patroni_barman` application
This commit adds the `config-switch` sub-command to `patroni_barman` application.

That sub-command communicates with the `pg-backup-api` to request a remote
execution of `barman config-switch` command.

It's designed to be used as a `on_role_change` callback, so one would be able
to change the Barman configuration by applying a model to stream from the new
primary of the Patroni cluster, for example.

The sub-command handles both the apply and unapply operations exposed by
`barman config-switch`.

Unit tests updated accordingly.

Referehces: PAT-154.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>