Retry one time on Etcd3 auth error
But do it only in case if we didn't authenticate right before executing
a request. Previously retries only happened when the caller was executed
with `Retry.__call__()`, which is not the case for methods like
`set_failover_value()` or `set_config_value()`.

In addition to that fix issues with `Retry.ensure_deadline(0)`:
1. the return value was ignored
2. we don't have to set `Retry.deadline` attr, it is not used anywhere

Close https://github.com/zalando/patroni/issues/3023