Fix plain Postgres synchronous replication mode
Since `synchronous_mode` was introduced to Patroni, the plain Postgres
synchronous replication has been no longer working.

The issue occurs because `process_sync_replication` always resets the
value of `synchronous_standby_names` in Postgres when `synchronous_mode`
is disabled in Patroni.

This commit fixes that issue by setting the value of
`synchronous_standby_names` as configured by the user, if that is the case,
when `synchronous_mode` is disabled.

Closes: #3093
References: PAT-254.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>