Advance permanent slots for cascading nodes while in failsafe
Lets consider a following replication setup:
```
primary->standby1->standby2(replicatefrom: standby1)
```

In this case the `primary` will not create a physical replication slot
for standby2, because it is streaming from the `standby1`.

Things will look differently if we have the following dynamic
configuration:
```yaml
slots:
    primary:
        type: physical
    standby1:
        type: physical
    standby2:
        type: physical
```

In this case `primary` will also have `standby2` physical replication
slot, which periodically must be advanced. So far it was working by
taking value of `xlog_location` from the `/members/standby2` key in DCS.

But, when DCS is down and failsafe mode is activate, the `standby2`
physical slot on the `primary` will not not be moved, because there was
not way to get the latest value of `xlog_location`.

This PR is addressing the problem by making replica nodes to return
their `xlog_location` as `lsn` header in the response on
`POST /failsafe` REST API request. The current primary will use these
values to advance replication slots for nodes with `nofailover` tag.