Get rid of SLOT_ADVANCE_AVAILABLE_VERSION in `dcs/__init__.py`
This constant was imported in `postgresql/__init__.py` and used in the
`can_advance_slots` property. But, after refactoring in #2958 we pass
around a reference to `Postgresql` instead of `major_version` and
therefore we can just rely on `can_advance_slots` property and don't
reimplement its logic in other places.