Configurable retention of members replication slots
Current problem of Patroni that strikes many people is that it removes
replication slot for member which key is expired from DCS. As a result,
when the replica comes back from a scheduled maintenance WAL segments
could be already absent, and it can't continue streaming without pulling
files from archive.
With PostgreSQL 16 and newer we get another problem: logical slot on a
standby node could be invalidated if physical replication slot on the
primary was removed (and `pg_catalog` vacuumed).
The most problematic environment is Kubernetes, where slot is removed
nearly instantly when member Pod is deleted.

So far, one of the recommended solutions was to configure permanent
physical slots with names that match member names to avoid removal of
replication slots. It works, but depending on environment might be
non-trivial to implement (when for example members may change their
names).

This PR implements support of `member_slots_ttl` global configuration
parameter, that controls for how long member replication slots should be
kept when the member key is absent. Default value is set to `30min`.
The feature is supported only starting from PostgreSQL 11 and newer,
because we want to retain slots not only on the leader node, but on all
nodes that could potentially become the new leader, and they should be
moved forward using `pg_replication_slot_advance()` function.

One could disable feature and get back to the old behavior by setting
`member_slots_ttl` to `0`.