Don't let the current node be chosen as synchronous
It could happen that there is "something" streaming from the current
primary node with `application_name` that matches name of the current
primary, for instance due to a faulty configuration.
When processing `pg_stat_replication` we only checked that the
`application_name` matches with the name one of the member nodes, but we
forgot to exclude our own name.
As a result there were following side-effects:
1. The current primary could be declared as a synchronous node.
2. As a result of [1] it wasn't possible to do a switchover.
3. During shutdown the current primary was waiting for itself to release
   it from synchronous nodes.

Close #3111