Make sure inactive hot physical replication slots don't hold xmin
Since `3.2.0` Patroni is able to create physical replication slots on
replica nodes just for the case if this node at some moment will become
the primary.
There are two potential problems of having such slots:
1. They prevent recycling of WAL files
2. They may affect vacuum on the primary is hot_standby_feedback is enabled.

The first class of issues is already addressed by periodically calling
pg_replication_slot_advance() function.
However the second class of issues doesn't happen instantly, but only
when the old primary switched to a replica. In this case physical
replication slots that were at some moment activate will hold NOT NULL
value of `xmin`, which will be propagated to the primary via
hot_standby_feedback mechanism.

To address the second problem we will detect that a physical replication
slot is not supposed to be active, but having NOT NULL `xmin` and
drop/crecreate it.

Close #3146