Recheck annotations when reading leader object on 409
There are cases when we may send the same PATCH request more than one
time to K8s API server and it could happen that the first request
actually successfully updated the target and we cancelled while waiting
for a response. The second PATCH request in this case will fail due to
resource_version mismatch.

So far our strategy for update_leader() method was - re-read the object
and repeat the request with the new resource_version. However, we can
avoid the update by comparing annotations on the read object with
annotations that we wanted to set.