Reduce log level of watchdog configuration failure
When in automatic mode we probably don't need to warn user about failure
to set up watchdog. This is the common case and makes many users think
that this feature is somehow necessary to run Patroni safely. For most
users it is completely fine to run without and it makes sense to reduce
their log spam.