Avoid dropping physical slots too early
Consider a situation: there is a permanent logical slot and primary and
replica are temporary down.
When Patroni is started on the former primary it starts Postgres in a
standby mode, what leads to removal of physical replication slot for the
replica because it has xmin.

We should postpone removal of such physical slots:
- on replica until there will be a leader in the cluster
- on primary until Postgres is promoted