Readiness API checks for replication on replicas
Readiness probes on Kubernetes are used for a few different things. One
is for determining pod disruption budget. Previous implementation
considered replicas ready as soon as PostgreSQL was started. This could
cause issues with async replication and rolling upgrades, where the
primary would be shut down before replica had even the time to start
replicating. Second important use is for determining which pods get
included as endpoints in a service. For both cases we would like to see
the pod only be considered ready when it is replicating and not too far
behind.

Also changes readiness to consider replicas as ready when failsafe is
active. In that case lag is ignored.