Maintain New Invariant For synchronous_mode_strict
This patch adds a new DCS update rule to maintain at least one
synchronous replica in /sync. Prevent updates with an empty set
by keeping the last known synchronous replica. This increases the
chances of being able to elect a new leader. For example, if there
are three nodes A, B and C, where A is the leader then the following
sequence prevents electing a new leader:

1. Shutdown B
2. Shutdown C
3. Shutdown A
4. Start B
5. Start C

If /sync is set '*' after step #2 then a new leader cannot be
elected when B and C are up despite one of them being a synchronous
replica. This patch skips updating the DCS such that sync_standby
will still contain C, which will allow the cluster to recover
correctly by promoting node C if A never joins back.