Make sure Etcd3 KVCache is not stale when doing get_cluster() (#3373)
We do it in a similar same way as in #3318 and in addition to that we cross check raft_term between `KVCache` and the client connection in the `is_ready()` method. In case of mismatch we reset `KVCache` state by terminating watcher connection.

Close https://github.com/patroni/patroni/issues/3359