Drop slots which have no restart_lsn
Instead of warning the user about the actions they might want to take,
this commit makes Patroni take the actions by itself. A slot without
restart_lsn has no value at all as it holds no WALs in the node.

We still issue a warning, but not we also take action on behalf of
the user.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>