Add a helper function to drop a physical slot
The same piece of code was kinda repeated 3 times in the `slots` module.
With that in mind, a helper function `_drop_physical_slot` was created
to handle that shared logic, reducing repeatition.

Appropriate unit tests will come through the next commit. This commit
focused simply on fixing the existing tests.

References: PAT-273.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>