Add dedicated unit test for `_drop_physical_slot`
This commit adds a dedicated test for checking the outcomes of the
new helper method `_drop_physical_slot`.

Besides that, it replaces most of the patches to `drop_replication_slot`
in `test_slots.py` with `_drop_physical_slot`. Only a couple of patches
were not replaced because the tests are basically checking the warning
messages which are issued by the underlying call to `drop_replication_slot`.
We need to check if it's worth keeping those or not.

References: PAR-273.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>