Handle case when HTTPConnection socket is wrapped with pyopenssl (#3425)
It happens when `pyopenssl` is installed, because `python-etcd` module enforces its usage. In this case we need to use `WrappedSocket.socket` to shutdown and close connection.
Not doing so results in an error:
```python
TypeError('WrappedSocket.shutdown() takes 1 positional argument but 2 were given')
```

Close https://github.com/patroni/patroni/issues/3420