Introduce Jepsen tests
Actual Clojure implementation is borrowed from
https://github.com/yandex/pgconsul/tree/main/docker/jepsen/jepsen

For tests we start Etcd v3 cluster consisting of 3 nodes and Patroni
cluster consisting of 3 nodes and configured to use quorum strict
replication.
Everything is started in privileged docker containers using
docker-compose.yml file. In addition to Etcd and Patroni in every
container we run sshd, because Jepsen is relying on it to create
iptables rules to emulate partitioned networks, to kill Patroni/Postgres
processes, and to run `patronictl switchover`.