Add support for custom `pg_rewind` options
This commit enhances the `pg_rewind` functionality to support passing
arbitrary command-line options from the Patroni configuration file.

It now reads the `postgresql.config.rewind` section (if present) and
uses the recently refactored `process_user_options` helper to parse
them. These options are then appended to the `pg_rewind` command line.

Besides allowing users to customize the `pg_rewind` behavior, this is
a necessary feature for supporting advanced functionality like TDE in
custom distributions of Postgres like EPAS, which require passing
custom flags to `pg_rewind`. A list of disallowed options is included
to prevent conflicts with parameters that are managed by Patroni, or
that would lead to an unexpected behavior of `pg_rewind`.

References: PAT-284.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>