Document custom `postgresql.rewind` options
This commit updates `yaml_configuration.rst` to add the new
`postgresql.rewind` configuration parameter.

It details how to pass custom options (as a list of strings and/or
key-value dictionaries) to the `pg_rewind` command and provides a
clear YAML example. It also lists the specific options that are not
allowed, as they are managed internally by Patroni.

References: PAT-284.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>