Improve reliability of _is_postmaster_process() checks (#3516)
When parsing `postmaster.pid` to determine whether a PostgreSQL cluster
is running for a given `$PGDATA`, we previously relied solely on
comparing the start time recorded in `postmaster.pid` with the process
creation time reported by the OS.

In some cases, the difference between these timestamps can exceed 3
seconds, leading to false negatives: PostgreSQL is running, but Patroni
incorrectly assumes it is not and attempts to start a second instance.

The improved check additionally verifies the process executable reported
by the OS (it must be postgres or postmaster) and the process current
working directory, which must match $PGDATA. This makes the detection
more robust and prevents false negatives.

Close https://github.com/patroni/patroni/issues/2792